+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [Unreleased]
+- Add in extra parameter for s3 download function for s3.download_fileobj ExtraArgs
 
 ## [1.4.0] - 2024-09-24
 - **CUMULUS-3838**